Merge

+++ FictionalFoodFinder/src/main/java/net/digitalstars/repository/ItemRepository.java
@@ -2,12 +2,11 @@ package net.digitalstars.repository;
 
 import java.util.List;
 import net.digitalstars.model.Item;
-import net.digitalstars.model.Truck;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository("itemRepository")
-public interface ItemRepository extends JpaRepository<Item, Integer>{
+public interface ItemRepository extends JpaRepository<Item, Long>{
     List<Item> findByTruckName(String truck_name);
     
 }//ItemRepository
